requirements.txt
flask
gunicorn
Flask-Failsafe
Pillow
opencv-python
mysql-connector-python
cryptography